extract default blueprint objects and functions handling them to its own file
